removed AliasConstants class and moved its content to AppDatabase class
