+++ app/src/main/java/org/eyeseetea/malariacare/database/model/Value.java
@@ -19,10 +19,10 @@
 
 package org.eyeseetea.malariacare.database.model;
 
-import static org.eyeseetea.malariacare.database.utils.AliasConstants.questionAlias;
-import static org.eyeseetea.malariacare.database.utils.AliasConstants.questionName;
-import static org.eyeseetea.malariacare.database.utils.AliasConstants.valueAlias;
-import static org.eyeseetea.malariacare.database.utils.AliasConstants.valueName;
+import static org.eyeseetea.malariacare.database.AppDatabase.questionAlias;
+import static org.eyeseetea.malariacare.database.AppDatabase.questionName;
+import static org.eyeseetea.malariacare.database.AppDatabase.valueAlias;
+import static org.eyeseetea.malariacare.database.AppDatabase.valueName;
 
 import com.raizlabs.android.dbflow.annotation.Column;
 import com.raizlabs.android.dbflow.annotation.PrimaryKey;
